Laravel 5+ cache .
Use built in cache of laravel 5 framework. AS the Laravel framework and Geotools use the same backends for caching  might as well just use one instance . eg: if Im using Redis for Laravel Cache and Geotools cache by using the frameworks cache for geotools I only need to manage cache settings in one place and easy to change cache backend's and have the framework and GeoTools updated to new backend.