[BUG] fix the polygon construct
The first test in the construct function is checking if the coordinates parameter is an array or null
then create a new CoordinateCollection
but a later test is using the set function when the parameter is an array then erases this new Coordinates collection.
So i rewamped the function keep thes argument tests but without the useless Coordinatecollection creation.